Add support for ADM in raw2bmx wave -> wave

+++ include/bmx/wave/WaveChunk.h
@@ -44,7 +44,7 @@ namespace bmx
 
 typedef mxfRIFFChunkIDType WaveChunkTag;
 
-#define WAVE_CHUNK_TAG(cstr)    (WaveChunkTag){(cstr)[0], (cstr)[1], (cstr)[2], (cstr)[3]}
+#define WAVE_CHUNK_TAG(cstr)    (WaveChunkTag){(uint8_t)((cstr)[0]), (uint8_t)((cstr)[1]), (uint8_t)((cstr)[2]), (uint8_t)((cstr)[3])}
 
 std::string get_wave_chunk_tag_str(WaveChunkTag tag);
